Separate head/tail notes from hold note class
